requirements.txt
# Core dependencies for Market Analysis Trading System
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
PyYAML>=6.0
requests>=2.31.0

# Optional: For better performance
scipy>=1.11.0

